Rule of threes.
