+++ packages/launcher/src/index.tsx
@@ -24,7 +24,7 @@ import {
 import * as vdom from '@phosphor/virtualdom';
 
 import {
-  Dialog, showDialog, VDomModel, VDomRenderer
+  showErrorMessage, VDomModel, VDomRenderer
 } from '@jupyterlab/apputils';
 
 import '../style/index.css';
@@ -399,7 +399,7 @@ function Card(kernel: boolean, item: ILauncherItem, launcher: Launcher, launcher
       launcher.dispose();
     }).catch(err => {
       launcher.pending = false;
-      Private.showErrorMessage('Launcher Error', err);
+      showErrorMessage('Launcher Error', err);
     });
   };
   // Add a data attribute for the category
@@ -445,21 +445,6 @@ namespace Private {
       iconLabel: options.iconLabel || '',
       rank: options.rank !== undefined ? options.rank : Infinity
     };
-  }
-
-  /**
-   * Show an error message dialog.
-   */
-  export
-  function showErrorMessage(title: string, error: any): Promise<void> {
-    console.error(error);
-    let options = {
-      title: title,
-      body: error.message || String(error),
-      buttons: [Dialog.okButton()],
-      okText: 'DISMISS'
-    };
-    return showDialog(options).then(() => { /* no-op */ });
   }
 
   /**
